test: improve performance by reusing commander instance
