+++ bin/aecli-crypto.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 /*
  * ISC License (ISC)
- * Copyright (c) 2018 aeternity developers
+ * Copyright (c) 2021 aeternity developers
  *
  *  Permission to use, copy, modify, and/or distribute this software for any
  *  purpose with or without fee is hereby granted, provided that the above
@@ -15,123 +15,7 @@
  *  OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
  *  PERFORMANCE OF THIS SOFTWARE.
  */
+const requireEsm = require('esm')(module)
+const genProgram = requireEsm('./commands/crypto')
 
-const program = require('commander')
-const fs = require('fs')
-const path = require('path')
-
-const requireEsm = require('esm')(module/*, options */) // use to handle es6 import/export
-const { prompt, PROMPT_TYPE } = requireEsm('./utils/prompt')
-const { Crypto } = requireEsm('@aeternity/aepp-sdk')
-const utils = requireEsm('./utils/index')
-
-program
-  .name('aecli crypto')
-
-// ## Key Extraction (from node nodes)
-async function extractReadableKeys (dir, options) {
-  const pwd = options.input
-  const password = await prompt(PROMPT_TYPE.askPassword)
-  const key = fs.readFileSync(path.join(pwd, dir, 'sign_key'))
-  const pubKey = fs.readFileSync(path.join(pwd, dir, 'sign_key.pub'))
-
-  const decrypted = Crypto.decryptPrivateKey(password, key)
-
-  const privateHex = Buffer.from(decrypted).toString('hex')
-  const decryptedPub = Crypto.decryptPubKey(password, pubKey)
-
-  console.log(`Private key (hex): ${privateHex}`)
-  console.log(`Public key (base check): ak_${Crypto.encodeBase58Check(decryptedPub)}`)
-  console.log(`Public key (hex): ${decryptedPub.toString('hex')}`)
-}
-
-// ## Key Pair Generation
-async function generateKeyPair (name, { output, password } = {}) {
-  await utils.account.generateSecureWallet(name, { output, password })
-}
-
-// ## Transaction Signing
-//
-// This function shows how to use a compliant private key to sign an æternity
-// transaction and turn it into an RLP-encoded tuple ready for mining
-function signTx (tx, privKey) {
-  if (!tx.match(/^tx_.+/)) {
-    throw Error('Not a valid transaction')
-  }
-
-  const binaryKey = (() => {
-    if (program.file) {
-      return fs.readFileSync(program.file)
-    } else if (privKey) {
-      return Buffer.from(privKey, 'hex')
-    } else {
-      throw Error('Must provide either [privkey] or [file]')
-    }
-  })()
-
-  const decryptedKey = program.password ? Crypto.decryptKey(program.password, binaryKey) : binaryKey
-
-  // Split the base58Check part of the transaction
-  const base58CheckTx = tx.split('_')[1]
-  // ... and sign the binary create_contract transaction
-  const binaryTx = Crypto.decodeBase58Check(base58CheckTx)
-
-  const signature = Crypto.sign(Buffer.concat([Buffer.from(Crypto.NETWORK_ID), binaryTx]), decryptedKey)
-
-  // the signed tx deserializer expects a 4-tuple:
-  // <tag, version, signatures_array, binary_tx>
-  const unpackedSignedTx = [
-    Buffer.from([11]),
-    Buffer.from([1]),
-    [Buffer.from(signature)],
-    binaryTx
-  ]
-
-  console.log(Crypto.encodeTx(unpackedSignedTx))
-}
-
-// ## Transaction Deserialization
-//
-// This helper function deserialized the transaction `tx` and prints the result.
-function unpackTx (tx) {
-  const deserializedTx = Crypto.deserialize(Crypto.decodeTx(tx))
-  console.log(JSON.stringify(deserializedTx, undefined, 2))
-}
-
-// ## Address decoder
-//
-// This helper function decodes address(base58) to hex
-function decodeAddress (address) {
-  const decoded = Crypto.decodeBase58Check(address.split('_')[1]).toString('hex')
-  console.log(`Decoded address (hex): ${decoded}`)
-}
-
-program
-  .command('decode <base58address>')
-  .description('Decodes base58 address to hex')
-  .action(decodeAddress)
-
-program
-  .command('decrypt <directory>')
-  .description('Decrypts public and private key to readable formats for testing purposes')
-  .option('-i, --input [directory]', 'Directory where to look for keys', '.')
-  .action(async (dir, ...args) => await extractReadableKeys(dir, utils.cli.getCmdFromArguments(args)))
-
-program
-  .command('genkey <keyname>')
-  .description('Generate keypair')
-  .option('-o, --output [directory]', 'Output directory for the keys', '.')
-  .option('-p, --password [directory]', 'Password for keypair', '.')
-  .action(async (keyname, ...args) => await generateKeyPair(keyname, utils.cli.getCmdFromArguments(args)))
-
-program
-  .command('sign <tx> [privkey]')
-  .option('-p, --password [password]', 'password of the private key')
-  .option('-f, --file [file]', 'private key file')
-  .action(signTx)
-
-program
-  .command('unpack <tx>')
-  .action(unpackTx)
-
-program.parseAsync()
+genProgram().parseAsync()
