+++ test/cli/aens.test.js
@@ -19,17 +19,21 @@ import { Crypto } from '@aeternity/aepp-sdk'
 import { before, describe, it } from 'mocha'
 import {
   configure,
-  execute as exec,
+  executeProgram,
   genAccount,
   plan,
   randomString,
   ready,
   WALLET_NAME
 } from './index'
+import nameProgramFactory from '../../bin/commands/name'
+import inspectProgramFactory from '../../bin/commands/inspect'
+import accountProgramFactory from '../../bin/commands/account'
 
+const executeName = (args) => executeProgram(nameProgramFactory, args, { withNetworkId: true })
+const executeInspect = (args) => executeProgram(inspectProgramFactory, args)
+const executeAccount = (args) => executeProgram(accountProgramFactory, args)
 plan(10000000000000)
-
-const execute = (arg) => exec(arg, { withNetworkId: true })
 
 function randomName (length, namespace = '.chain') {
   return randomString(length).toLowerCase() + namespace
@@ -54,8 +58,7 @@ describe('CLI AENS Module', function () {
 
   it('Full claim', async () => {
     const updateTx = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'full-claim',
         WALLET_NAME,
         '--password',
@@ -73,8 +76,7 @@ describe('CLI AENS Module', function () {
 
   it('Full claim with options', async () => {
     const updateTx = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'full-claim',
         WALLET_NAME,
         '--password',
@@ -100,8 +102,7 @@ describe('CLI AENS Module', function () {
 
   it('Pre Claim Name', async () => {
     const preClaim = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'pre-claim',
         WALLET_NAME,
         '--password',
@@ -110,7 +111,7 @@ describe('CLI AENS Module', function () {
         '--json'
       ])
     )
-    const nameResult = JSON.parse(await exec(['inspect', name2, '--json']))
+    const nameResult = JSON.parse(await executeInspect([name2, '--json']))
     salt = preClaim.salt
 
     preClaim.blockHeight.should.be.gt(0)
@@ -122,8 +123,7 @@ describe('CLI AENS Module', function () {
 
   it('Claim Name', async () => {
     const claim = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'claim',
         WALLET_NAME,
         '--password',
@@ -133,7 +133,7 @@ describe('CLI AENS Module', function () {
         '--json'
       ])
     )
-    const nameResult = JSON.parse(await exec(['inspect', name2, '--json']))
+    const nameResult = JSON.parse(await executeInspect([name2, '--json']))
 
     claim.blockHeight.should.be.gt(0)
     claim.pointers.length.should.be.equal(0)
@@ -142,8 +142,7 @@ describe('CLI AENS Module', function () {
 
   it('Update Name', async () => {
     const updateTx = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'update',
         WALLET_NAME,
         name2,
@@ -153,7 +152,7 @@ describe('CLI AENS Module', function () {
         '--json'
       ])
     )
-    const nameResult = JSON.parse(await exec(['inspect', name2, '--json']))
+    const nameResult = JSON.parse(await executeInspect([name2, '--json']))
 
     updateTx.blockHeight.should.be.gt(0)
     const isUpdatedNode = !!nameResult.pointers.find(
@@ -166,8 +165,7 @@ describe('CLI AENS Module', function () {
   it('extend name ttl', async () => {
     const height = await wallet.height()
     const extendTx = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'extend',
         WALLET_NAME,
         name2,
@@ -178,7 +176,7 @@ describe('CLI AENS Module', function () {
       ])
     )
 
-    const nameResult = JSON.parse(await exec(['inspect', name2, '--json']))
+    const nameResult = JSON.parse(await executeInspect([name2, '--json']))
     const isExtended = nameResult.ttl - 50 >= height
     isExtended.should.be.equal(true)
     extendTx.blockHeight.should.be.gt(0)
@@ -187,8 +185,7 @@ describe('CLI AENS Module', function () {
 
   it('Fail spend by name on invalid input', async () => {
     const amount = 100000009
-    await execute([
-      'account',
+    await executeAccount([
       'spend',
       WALLET_NAME,
       '--password',
@@ -202,8 +199,7 @@ describe('CLI AENS Module', function () {
   it('Spend by name', async () => {
     const amount = 100000009
     const spendTx = JSON.parse(
-      await execute([
-        'account',
+      await executeAccount([
         'spend',
         WALLET_NAME,
         '--password',
@@ -225,8 +221,7 @@ describe('CLI AENS Module', function () {
     await wallet.addAccount(account)
 
     const transferTx = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'transfer',
         WALLET_NAME,
         name2,
@@ -248,8 +243,7 @@ describe('CLI AENS Module', function () {
 
   it('Revoke Name', async () => {
     const revoke = JSON.parse(
-      await execute([
-        'name',
+      await executeName([
         'revoke',
         WALLET_NAME,
         '--password',
@@ -259,7 +253,7 @@ describe('CLI AENS Module', function () {
       ])
     )
 
-    const nameResult = JSON.parse(await exec(['inspect', name2, '--json']))
+    const nameResult = JSON.parse(await executeInspect([name2, '--json']))
 
     revoke.blockHeight.should.be.gt(0)
     nameResult.status.should.equal('AVAILABLE')
@@ -283,8 +277,7 @@ describe('CLI AENS Module', function () {
 
     it('Make bid', async () => {
       const bid = JSON.parse(
-        await execute([
-          'name',
+        await executeName([
           'bid',
           WALLET_NAME,
           '--password',
@@ -301,8 +294,7 @@ describe('CLI AENS Module', function () {
 
     it('Fail on open  again', async () => {
       const preClaim = JSON.parse(
-        await execute([
-          'name',
+        await executeName([
           'pre-claim',
           WALLET_NAME,
           '--password',
@@ -311,8 +303,7 @@ describe('CLI AENS Module', function () {
           '--json'
         ])
       )
-      await execute([
-        'name',
+      await executeName([
         'claim',
         WALLET_NAME,
         '--password',
