+++ test/cli/chain.test.js
@@ -16,7 +16,10 @@
  */
 
 import { before, describe, it } from 'mocha'
-import { configure, BaseAe, execute, parseBlock, ready, genAccount } from './index'
+import { configure, BaseAe, executeProgram, parseBlock, ready, genAccount } from './index'
+import chainProgramFactory from '../../bin/commands/chain'
+
+const executeChain = args => executeProgram(chainProgramFactory, args)
 
 describe('CLI Chain Module', function () {
   let wallet
@@ -27,7 +30,7 @@ describe('CLI Chain Module', function () {
     wallet = await ready(this)
   })
   it('TOP', async () => {
-    const res = JSON.parse(await execute(['chain', 'top', '--json']))
+    const res = JSON.parse(await executeChain(['top', '--json']))
     res.should.be.a('object')
     res.hash.should.be.a('string')
     res.height.should.be.a('number')
@@ -37,11 +40,11 @@ describe('CLI Chain Module', function () {
     await wallet.addAccount(genAccount(), { select: true })
 
     const { nodeVersion } = await wallet.api.getStatus()
-    const res = JSON.parse(await execute(['chain', 'status', '--json']))
+    const res = JSON.parse(await executeChain(['status', '--json']))
     res.nodeVersion.should.equal(nodeVersion)
   })
   it('PLAY', async () => {
-    const res = await execute(['chain', 'play', '--limit', '4'])
+    const res = await executeChain(['play', '--limit', '4'])
     res.split('<<------------------------------------->>').length.should.equal(5)
 
     const parsed = res.split('<<------------------------------------->>').map(parseBlock)
@@ -51,7 +54,7 @@ describe('CLI Chain Module', function () {
   })
   it('TTL', async () => {
     const [res] = await Promise.all([
-      execute(['chain', 'ttl', 10, '--json'])
+      executeChain(['ttl', 10, '--json'])
     ])
     const height = await wallet.height()
     const { relativeTtl } = JSON.parse(res)
@@ -60,7 +63,7 @@ describe('CLI Chain Module', function () {
   })
   it('NETWORK ID', async () => {
     const nodeNetworkId = wallet.getNetworkId()
-    const { networkId } = JSON.parse((await execute(['chain', 'network_id', '--json'])))
+    const { networkId } = JSON.parse((await executeChain(['network_id', '--json'])))
     nodeNetworkId.should.equal(networkId)
   })
 })
