+++ test/cli/oracle.test.js
@@ -17,11 +17,11 @@
 
 import { Crypto } from '@aeternity/aepp-sdk'
 import { before, describe, it } from 'mocha'
-import { configure, execute as exec, plan, ready, WALLET_NAME } from './index'
+import { configure, executeProgram, plan, ready, WALLET_NAME } from './index'
+import oracleProgramFactory from '../../bin/commands/oracle'
 
+const executeOracle = args => executeProgram(oracleProgramFactory, args, { withNetworkId: true })
 plan(10000000000000)
-
-const execute = (arg) => exec(arg, { withNetworkId: true })
 
 describe('CLI Oracle Module', function () {
   configure(this)
@@ -37,8 +37,8 @@ describe('CLI Oracle Module', function () {
   })
 
   it('Oracle create', async () => {
-    const oracleCreate = JSON.parse(await execute([
-      'oracle', 'create', WALLET_NAME, '--password', 'test', oracleFormat, responseFormat, '--json'
+    const oracleCreate = JSON.parse(await executeOracle([
+      'create', WALLET_NAME, '--password', 'test', oracleFormat, responseFormat, '--json'
     ]))
     oracleCreate.blockHeight.should.be.gt(0)
     oracleCreate.queryFormat.should.be.equal(oracleFormat)
@@ -48,16 +48,16 @@ describe('CLI Oracle Module', function () {
 
   it('Oracle extend', async () => {
     const oracle = await wallet.getOracleObject(oracleId)
-    const oracleExtend = JSON.parse(await execute([
-      'oracle', 'extend', WALLET_NAME, '--password', 'test', oracleId, 100, '--json'
+    const oracleExtend = JSON.parse(await executeOracle([
+      'extend', WALLET_NAME, '--password', 'test', oracleId, 100, '--json'
     ]))
     oracleExtend.blockHeight.should.be.gt(0)
     oracleExtend.ttl.should.be.gte(oracle.ttl + 100)
   })
 
   it('Oracle create query', async () => {
-    const oracleQuery = JSON.parse(await execute([
-      'oracle', 'create-query', WALLET_NAME, '--password', 'test', oracleId, 'Hello?', '--json'
+    const oracleQuery = JSON.parse(await executeOracle([
+      'create-query', WALLET_NAME, '--password', 'test', oracleId, 'Hello?', '--json'
     ]))
     oracleQuery.blockHeight.should.be.gt(0)
     oracleQuery.decodedQuery.should.be.equal('Hello?')
@@ -68,8 +68,8 @@ describe('CLI Oracle Module', function () {
   })
 
   it('Oracle respond to query', async () => {
-    const oracleQueryResponse = JSON.parse(await execute([
-      'oracle', 'respond-query', WALLET_NAME, '--password', 'test', oracleId, queryId, 'Hi!', '--json'
+    const oracleQueryResponse = JSON.parse(await executeOracle([
+      'respond-query', WALLET_NAME, '--password', 'test', oracleId, queryId, 'Hi!', '--json'
     ]))
     oracleQueryResponse.blockHeight.should.be.gt(0)
     const oracle = await wallet.getOracleObject(oracleId)
@@ -79,14 +79,14 @@ describe('CLI Oracle Module', function () {
 
   it('Get non existed Oracle', async () => {
     const fakeOracleId = Crypto.generateKeyPair().publicKey.replace('ak_', 'ok_')
-    await execute(['oracle', 'get', fakeOracleId, '--json'])
+    await executeOracle(['get', fakeOracleId, '--json'])
       .should.be.rejectedWith('error: Oracle not found')
-    await execute(['oracle', 'get', 'oq_d1sadasdasda', '--json'])
+    await executeOracle(['get', 'oq_d1sadasdasda', '--json'])
       .should.be.rejectedWith('Encoded string have a wrong type: oq (expected: ok)')
   })
 
   it('Get existed Oracle', async () => {
-    const oracle = JSON.parse(await execute(['oracle', 'get', oracleId, '--json']))
+    const oracle = JSON.parse(await executeOracle(['get', oracleId, '--json']))
     oracle.id.should.be.a('string')
     oracle.id.split('_')[0].should.be.equal('ok')
   })
