+++ test/cli/tx.js
@@ -18,7 +18,12 @@
 import { Crypto, MemoryAccount } from '@aeternity/aepp-sdk'
 import fs from 'fs'
 import { after, before, describe, it } from 'mocha'
-import { BaseAe, configure, execute, parseBlock, randomString, ready } from './index'
+import { BaseAe, configure, executeProgram, parseBlock, randomString, ready } from './index'
+import txProgramFactory from '../../bin/commands/tx'
+import accountProgramFactory from '../../bin/commands/account'
+import chainProgramFactory from '../../bin/commands/chain'
+
+const executeTx = (...args) => executeProgram(txProgramFactory, ...args)
 
 const WALLET_NAME = 'txWallet'
 const testContract = `
@@ -33,10 +38,10 @@ function randomName (length = 18, namespace = '.chain') {
 }
 
 async function signAndPost (tx, assert) {
-  const { signedTx } = JSON.parse(await execute(['account', 'sign', WALLET_NAME, tx, '--password', 'test', '--json'], { withNetworkId: true }))
+  const { signedTx } = JSON.parse(await executeProgram(accountProgramFactory, ['sign', WALLET_NAME, tx, '--password', 'test', '--json'], { withNetworkId: true }))
   return assert
-    ? (await execute(['chain', 'broadcast', signedTx, '--no-waitMined'])).should.contain('Transaction send to the chain')
-    : execute(['chain', 'broadcast', signedTx])
+    ? (await executeProgram(chainProgramFactory, ['broadcast', signedTx, '--no-waitMined'])).should.contain('Transaction send to the chain')
+    : executeProgram(chainProgramFactory, ['broadcast', signedTx])
 }
 
 describe('CLI Transaction Module', function () {
@@ -56,7 +61,7 @@ describe('CLI Transaction Module', function () {
     compilerCLI = await ready(this)
     const GENESIS = await BaseAe()
     await GENESIS.spend('100000000000000000000000000', TX_KEYS.publicKey)
-    await execute(['account', 'save', WALLET_NAME, '--password', 'test', TX_KEYS.secretKey, '--overwrite'])
+    await executeProgram(accountProgramFactory, ['save', WALLET_NAME, '--password', 'test', TX_KEYS.secretKey, '--overwrite'])
     wallet = await BaseAe()
     await wallet.addAccount(MemoryAccount({ keypair: TX_KEYS }))
     fs.writeFileSync('contractTest', testContract)
@@ -70,13 +75,13 @@ describe('CLI Transaction Module', function () {
   it('Build spend tx offline and send on-chain', async () => {
     const amount = 100
 
-    const { tx } = JSON.parse(await execute(['tx', 'spend', TX_KEYS.publicKey, TX_KEYS.publicKey, amount, nonce, '--json']))
+    const { tx } = JSON.parse(await executeTx(['spend', TX_KEYS.publicKey, TX_KEYS.publicKey, amount, nonce, '--json']))
     await signAndPost(tx, true)
     nonce += 1
   })
 
   it('Build preclaim tx offline and send on-chain', async () => {
-    const { tx, salt: nameSalt } = JSON.parse(await execute(['tx', 'name-preclaim', TX_KEYS.publicKey, name, nonce, '--json']))
+    const { tx, salt: nameSalt } = JSON.parse(await executeTx(['name-preclaim', TX_KEYS.publicKey, name, nonce, '--json']))
     salt = nameSalt
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
@@ -85,7 +90,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build claim tx offline and send on-chain', async () => {
-    const { tx } = JSON.parse(await execute(['tx', 'name-claim', TX_KEYS.publicKey, salt, name, nonce, '--json']))
+    const { tx } = JSON.parse(await executeTx(['name-claim', TX_KEYS.publicKey, salt, name, nonce, '--json']))
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
     isMined.should.be.equal(true)
@@ -95,7 +100,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build update tx offline and send on-chain', async () => {
-    const { tx } = JSON.parse(await execute(['tx', 'name-update', TX_KEYS.publicKey, nameId, nonce, TX_KEYS.publicKey, '--json']))
+    const { tx } = JSON.parse(await executeTx(['name-update', TX_KEYS.publicKey, nameId, nonce, TX_KEYS.publicKey, '--json']))
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
     isMined.should.be.equal(true)
@@ -103,7 +108,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build transfer tx offline and send on-chain', async () => {
-    const { tx } = JSON.parse(await execute(['tx', 'name-transfer', TX_KEYS.publicKey, TX_KEYS.publicKey, nameId, nonce, '--json']))
+    const { tx } = JSON.parse(await executeTx(['name-transfer', TX_KEYS.publicKey, TX_KEYS.publicKey, nameId, nonce, '--json']))
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
     isMined.should.be.equal(true)
@@ -111,7 +116,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build revoke tx offline and send on-chain', async () => {
-    const { tx } = JSON.parse(await execute(['tx', 'name-revoke', TX_KEYS.publicKey, nameId, nonce, '--json']))
+    const { tx } = JSON.parse(await executeTx(['name-revoke', TX_KEYS.publicKey, nameId, nonce, '--json']))
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
     isMined.should.be.equal(true)
@@ -121,7 +126,7 @@ describe('CLI Transaction Module', function () {
   it('Build contract create tx offline and send on-chain', async () => {
     const { bytecode } = await compilerCLI.contractCompile(testContract)
     const callData = await compilerCLI.contractEncodeCallDataAPI(testContract, 'init', [])
-    const { tx, contractId: cId } = JSON.parse(await execute(['tx', 'contract-deploy', TX_KEYS.publicKey, bytecode, callData, nonce, '--json']))
+    const { tx, contractId: cId } = JSON.parse(await executeTx(['contract-deploy', TX_KEYS.publicKey, bytecode, callData, nonce, '--json']))
     contractId = cId
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
@@ -131,7 +136,7 @@ describe('CLI Transaction Module', function () {
 
   it('Build contract call tx offline and send on-chain', async () => {
     const callData = await compilerCLI.contractEncodeCallDataAPI(testContract, 'test', ['1', '2'])
-    const { tx } = JSON.parse(await execute(['tx', 'contract-call', TX_KEYS.publicKey, contractId, callData, nonce, '--json']))
+    const { tx } = JSON.parse(await executeTx(['contract-call', TX_KEYS.publicKey, contractId, callData, nonce, '--json']))
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
     isMined.should.be.equal(true)
@@ -139,7 +144,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build oracle register tx offline and send on-chain', async () => {
-    const result = await execute(['tx', 'oracle-register', TX_KEYS.publicKey, '{city: "str"}', '{tmp:""num}', nonce, '--json'], { withOutReject: true })
+    const result = await executeTx(['oracle-register', TX_KEYS.publicKey, '{city: "str"}', '{tmp:""num}', nonce, '--json'], { withOutReject: true })
     const { tx } = JSON.parse(result)
     const res = (parseBlock(await signAndPost(tx)))
     const isMined = !isNaN(res.block_height_)
@@ -149,7 +154,7 @@ describe('CLI Transaction Module', function () {
 
   it('Build oracle extend  tx offline and send on-chain', async () => {
     const oracleCurrentTtl = await wallet.api.getOracleByPubkey(oracleId)
-    const { tx } = JSON.parse(await execute(['tx', 'oracle-extend', TX_KEYS.publicKey, oracleId, 100, nonce, '--json'], { withOutReject: true }))
+    const { tx } = JSON.parse(await executeTx(['oracle-extend', TX_KEYS.publicKey, oracleId, 100, nonce, '--json'], { withOutReject: true }))
     const res = (parseBlock(await signAndPost(tx)))
     const oracleTtl = await wallet.api.getOracleByPubkey(oracleId)
     const isExtended = +oracleTtl.ttl === +oracleCurrentTtl.ttl + 100
@@ -160,7 +165,7 @@ describe('CLI Transaction Module', function () {
   })
 
   it('Build oracle post query tx offline and send on-chain', async () => {
-    const { tx } = JSON.parse(await execute(['tx', 'oracle-post-query', TX_KEYS.publicKey, oracleId, '{city: "Berlin"}', nonce, '--json'], { withOutReject: true }))
+    const { tx } = JSON.parse(await executeTx(['oracle-post-query', TX_KEYS.publicKey, oracleId, '{city: "Berlin"}', nonce, '--json'], { withOutReject: true }))
     const res = (parseBlock(await signAndPost(tx)))
     const { oracleQueries: queries } = await wallet.api.getOracleQueriesByPubkey(oracleId)
     queryId = queries[0].id
@@ -173,7 +178,7 @@ describe('CLI Transaction Module', function () {
 
   it('Build oracle respond tx offline and send on-chain', async () => {
     const response = '{tmp: 10}'
-    const { tx } = JSON.parse(await execute(['tx', 'oracle-respond', TX_KEYS.publicKey, oracleId, queryId, response, nonce, '--json'], { withOutReject: true }))
+    const { tx } = JSON.parse(await executeTx(['oracle-respond', TX_KEYS.publicKey, oracleId, queryId, response, nonce, '--json'], { withOutReject: true }))
     const res = (parseBlock(await signAndPost(tx)))
     const { oracleQueries: queries } = await wallet.api.getOracleQueriesByPubkey(oracleId)
     const responseQuery = Crypto.decodeBase64Check(queries[0].response.slice(3)).toString()
